requirements.txt
Django==4.1.6
djangorestframework==3.14.0
django-environ==0.9.0
django-cors-headers==3.13.0